:truck: Rename chain module to chain_module and add to add_module
