+++ thinc/layers/add_module.py
