+++ thinc/layers/chain_module.py
